Recreate and move full-stack example to dedicated directory
