--- nflow-examples/spring-boot/full-stack/build.gradle
@@ -1,37 +1,0 @@
-plugins {
-  id 'de.undercouch.download' version '4.0.2'
-  id "com.github.ben-manes.versions" version "0.27.0"
-  id 'org.springframework.boot' version '2.2.1.RELEASE'
-  id 'io.spring.dependency-management' version '1.0.8.RELEASE'
-}
-
-apply plugin: 'java'
-apply plugin: 'eclipse'
-
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = 1.8
-
-repositories {
-  mavenCentral()
-}
-
-configurations {
-  nflowExplorer
-}
-
-dependencies {
-  implementation 'org.springframework.boot:spring-boot-starter-web'
-  implementation 'org.springframework.boot:spring-boot-starter-jersey'
-  implementation 'org.springframework.boot:spring-boot-starter-jdbc'
-  implementation 'io.nflow:nflow-rest-api-jax-rs:7.2.2'
-  runtimeOnly 'com.h2database:h2'
-  testImplementation 'org.springframework.boot:spring-boot-starter-test'
-  nflowExplorer group: 'io.nflow', name: 'nflow-explorer', version: '7.2.2', ext: 'tar.gz'
-}
-
-task resolveNflowExplorer(type: Copy) {
-  destinationDir = file("$buildDir/resources/main/static/explorer")
-  from { tarTree(resources.gzip(configurations.nflowExplorer.singleFile)) }
-}
-
-processResources.dependsOn resolveNflowExplorer
