--- nflow-examples/spring-boot/full-stack/src/main/resources/application.properties
@@ -1,4 +1,0 @@
-spring.profiles.active=nflow.db.h2
-logging.level.root=WARN
-spring.jersey.application-path=/rest
-nflow.db.h2.url=jdbc:h2:mem:test;TRACE_LEVEL_FILE=4;DB_CLOSE_ON_EXIT=FALSE
