--- nflow-examples/spring-boot/full-stack/src/test/java/io/nflow/springboot/fullstack/SpringBootFullStackApplicationTests.java
@@ -1,17 +1,0 @@
-package io.nflow.springboot.fullstack;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SpringBootFullStackApplicationTests {
-
-  @Test
-  public void applicationStarts() {
-    // just check that the application starts
-  }
-
-}
